+++ Sources/Sentry/SentryClient.m
@@ -7,6 +7,7 @@
 #import "SentryEnvelope.h"
 #import "SentryEvent.h"
 #import "SentryFileManager.h"
+#import "SentryFrameRemover.h"
 #import "SentryGlobalEventProcessor.h"
 #import "SentryId.h"
 #import "SentryLog.h"
@@ -47,7 +48,9 @@ SentryClient ()
         self.debugMetaBuilder =
             [[SentryDebugMetaBuilder alloc] initWithBinaryImageProvider:provider];
 
-        SentryStacktraceBuilder *stacktraceBuilder = [[SentryStacktraceBuilder alloc] init];
+        SentryFrameRemover *frameRemover = [[SentryFrameRemover alloc] init];
+        SentryStacktraceBuilder *stacktraceBuilder =
+            [[SentryStacktraceBuilder alloc] initWithSentryFrameRemover:frameRemover];
         id<SentryCrashMachineContextWrapper> machineContextWrapper =
             [[SentryCrashDefaultMachineContextWrapper alloc] init];
 
@@ -276,9 +279,7 @@ SentryClient ()
 
     BOOL threadsNotAttached = !(nil != event.threads && event.threads.count > 0);
     if (shouldAttachStacktrace && threadsNotAttached) {
-        // We don't want to add the stacktrace of attaching the stacktrace.
-        // Therefore we skip three frames.
-        event.threads = [self.threadInspector getCurrentThreadsSkippingFrames:3];
+        event.threads = [self.threadInspector getCurrentThreads];
     }
 
     if (nil != scope) {