+++ Tests/SentryTests/SentryClientTests.swift
@@ -12,7 +12,7 @@ class SentryClientTest: XCTestCase {
         )
         
         let threadInspector = SentryThreadInspector(
-            stacktraceBuilder: SentryStacktraceBuilder(),
+            stacktraceBuilder: SentryStacktraceBuilder(sentryFrameRemover: SentryFrameRemover()),
             andMachineContextWrapper: SentryCrashDefaultMachineContextWrapper()
         )
         
@@ -209,7 +209,7 @@ class SentryClientTest: XCTestCase {
         
         eventId.assertIsNotEmpty()
         assertLastSentEvent { actual in
-            assertValidErrorEvent(actual, framesToSkip: 6)
+            assertValidErrorEvent(actual)
         }
     }
     
@@ -267,7 +267,7 @@ class SentryClientTest: XCTestCase {
         
         eventId.assertIsNotEmpty()
         assertLastSentEvent { actual in
-            assertValidExceptionEvent(actual, framesToSkip: 6)
+            assertValidExceptionEvent(actual)
         }
     }
     
@@ -364,7 +364,7 @@ class SentryClientTest: XCTestCase {
     func testNoDsn_MessageNotSent() {
         let sut = fixture.getSutWithNoDsn()
         let eventId = sut.capture(message: message, scope: nil)
-        
+
         eventId.assertIsEmpty()
         assertNothingSent()
     }
@@ -372,7 +372,7 @@ class SentryClientTest: XCTestCase {
     func testNoDsn_ExceptionNotSent() {
         let sut = fixture.getSutWithNoDsn()
         let eventId = sut.capture(exception: exception, scope: nil)
-        
+
         eventId.assertIsEmpty()
         assertNothingSent()
     }
@@ -380,7 +380,7 @@ class SentryClientTest: XCTestCase {
     func testNoDsn_ErrorNotSent() {
         let sut = fixture.getSutWithNoDsn()
         let eventId = sut.capture(error: error, scope: nil)
-        
+
         eventId.assertIsEmpty()
         assertNothingSent()
     }
@@ -390,36 +390,36 @@ class SentryClientTest: XCTestCase {
         fixture.getSut(configureOptions: { options in
             options.dsn = nil
         }).capture(session: fixture.session)
-        
+
         assertNothingSent()
     }
-    
+
     func testNoDsn_EventWithSessionsNotSent() {
         _ = SentryEnvelope(event: Event())
         let eventId = fixture.getSut(configureOptions: { options in
             options.dsn = nil
         }).capture(Event(), with: fixture.session, with: nil)
-        
+
         eventId.assertIsEmpty()
         assertNothingSent()
     }
-    
+
     func testNoDsn_ExceptionWithSessionsNotSent() {
         _ = SentryEnvelope(event: Event())
         let eventId = fixture.getSut(configureOptions: { options in
             options.dsn = nil
         }).capture(self.exception, with: fixture.session, with: nil)
-        
+
         eventId.assertIsEmpty()
         assertNothingSent()
     }
-    
+
     func testNoDsn_ErrorWithSessionsNotSent() {
         _ = SentryEnvelope(event: Event())
         let eventId = fixture.getSut(configureOptions: { options in
             options.dsn = nil
         }).captureError(self.error, with: fixture.session, with: nil)
-        
+
         eventId.assertIsEmpty()
         assertNothingSent()
     }
@@ -496,18 +496,18 @@ class SentryClientTest: XCTestCase {
         }
     }
     
-    private func assertValidErrorEvent(_ event: Event, framesToSkip: Int = 5) {
+    private func assertValidErrorEvent(_ event: Event) {
         XCTAssertEqual(SentryLevel.error, event.level)
         XCTAssertEqual(error.localizedDescription, event.message)
         assertValidDebugMeta(actual: event.debugMeta)
-        assertValidThreads(actual: event.threads, framesToSkip)
+        assertValidThreads(actual: event.threads)
     }
     
-    private func assertValidExceptionEvent(_ event: Event, framesToSkip: Int = 5) {
+    private func assertValidExceptionEvent(_ event: Event) {
         XCTAssertEqual(SentryLevel.error, event.level)
         XCTAssertEqual(exception.reason, event.message)
         assertValidDebugMeta(actual: event.debugMeta)
-        assertValidThreads(actual: event.threads, framesToSkip)
+        assertValidThreads(actual: event.threads)
     }
     
     private func assertLastSentEnvelope(assert: (SentryEnvelope) -> Void) {
@@ -530,23 +530,27 @@ class SentryClientTest: XCTestCase {
         XCTAssertEqual(debugMetas, actual ?? [])
     }
     
-    private func assertValidThreads(actual: [Sentry.Thread]?, _ framesToSkip: Int = 5) {
+    private func assertValidThreads(actual: [Sentry.Thread]?) {
+        let expected = fixture.threadInspector.getCurrentThreads()
+
         // We can only compare the stacktrace up to the test method. Therefore we
-        // need to skip a few frames for the expected stacktrace and also two
-        // frame for the actual stacktrace.
-        let expected = fixture.threadInspector.getCurrentThreadsSkippingFrames(framesToSkip)
-        var actualFrames = actual?[0].stacktrace?.frames ?? []
-        XCTAssertTrue(actualFrames.count > 1, "Event has no stacktrace.")
-        if actualFrames.count > 1 {
-            actualFrames.remove(at: actualFrames.count - 1)
-            actualFrames.remove(at: actualFrames.count - 1)
-            actual?[0].stacktrace?.frames = actualFrames
-        }
+        // need to remove a few frames for the stacktraces.
+        removeFrames(threads: expected)
+        removeFrames(threads: actual ?? [])
         
         XCTAssertEqual(expected.count, actual?.count)
         XCTAssertEqual(expected, actual ?? [])
     }
-    
+
+    private func removeFrames(threads: [Sentry.Thread]) {
+        var actualFrames = threads[0].stacktrace?.frames ?? []
+        XCTAssertTrue(actualFrames.count > 1, "Event has no stacktrace.")
+        if actualFrames.count > 1 {
+            actualFrames.removeLast(3)
+            threads[0].stacktrace?.frames = actualFrames
+        }
+    }
+
     private func assertNothingSent() {
         XCTAssertNil(fixture.transport.lastSentEnvelope)
         XCTAssertEqual(0, fixture.transport.sentEventsWithSession.count)